+++ crates/rs-core/src/db/templates.rs
@@ -137,6 +137,7 @@ pub async fn get_template_endpoints(
             position_last: r.get("position_last"),
             delivered_bytes: r.get("delivered_bytes"),
             is_fast: r.get::<i32, _>("is_fast") != 0,
+            pusher: Default::default(),
             created_at: r.get("created_at"),
             updated_at: r.get("updated_at"),
         })
